+++ .oclif.manifest.json
@@ -1043,58 +1043,6 @@
         "tarballs.js"
       ]
     },
-    "setup:install": {
-      "aliases": [],
-      "args": {},
-      "description": "Automatically check and install system prerequisites",
-      "examples": [
-        "eggs setup                           # this help",
-        "sudo eggs setup install              # install native dependencies, autocomplete, man, etc",
-        "sudo eggs setup purge                # purge all configurations, autocomplete, man, etc installed from penguins-eggs AppImage"
-      ],
-      "flags": {},
-      "hasDynamicHelp": false,
-      "hiddenAliases": [],
-      "id": "setup:install",
-      "pluginAlias": "penguins-eggs",
-      "pluginName": "penguins-eggs",
-      "pluginType": "core",
-      "strict": true,
-      "enableJsonFlag": false,
-      "isESM": true,
-      "relativePath": [
-        "dist",
-        "commands",
-        "setup",
-        "install.js"
-      ]
-    },
-    "setup:purge": {
-      "aliases": [],
-      "args": {},
-      "description": "Automatically check and install system prerequisites",
-      "examples": [
-        "eggs setup                           # this help",
-        "sudo eggs setup install              # install native dependencies, autocomplete, man, etc",
-        "sudo eggs setup purge                # purge all configurations, autocomplete, man, etc installed from penguins-eggs AppImage"
-      ],
-      "flags": {},
-      "hasDynamicHelp": false,
-      "hiddenAliases": [],
-      "id": "setup:purge",
-      "pluginAlias": "penguins-eggs",
-      "pluginName": "penguins-eggs",
-      "pluginType": "core",
-      "strict": true,
-      "enableJsonFlag": false,
-      "isESM": true,
-      "relativePath": [
-        "dist",
-        "commands",
-        "setup",
-        "purge.js"
-      ]
-    },
     "tools:clean": {
       "aliases": [],
       "args": {},
@@ -1567,6 +1515,58 @@
         "commands",
         "wardrobe",
         "wear.js"
+      ]
+    },
+    "setup:install": {
+      "aliases": [],
+      "args": {},
+      "description": "Automatically check and install system prerequisites",
+      "examples": [
+        "eggs setup                           # this help",
+        "sudo eggs setup install              # install native dependencies, autocomplete, man, etc",
+        "sudo eggs setup purge                # purge all configurations, autocomplete, man, etc installed from penguins-eggs AppImage"
+      ],
+      "flags": {},
+      "hasDynamicHelp": false,
+      "hiddenAliases": [],
+      "id": "setup:install",
+      "pluginAlias": "penguins-eggs",
+      "pluginName": "penguins-eggs",
+      "pluginType": "core",
+      "strict": true,
+      "enableJsonFlag": false,
+      "isESM": true,
+      "relativePath": [
+        "dist",
+        "commands",
+        "setup",
+        "install.js"
+      ]
+    },
+    "setup:purge": {
+      "aliases": [],
+      "args": {},
+      "description": "Automatically check and install system prerequisites",
+      "examples": [
+        "eggs setup                           # this help",
+        "sudo eggs setup install              # install native dependencies, autocomplete, man, etc",
+        "sudo eggs setup purge                # purge all configurations, autocomplete, man, etc installed from penguins-eggs AppImage"
+      ],
+      "flags": {},
+      "hasDynamicHelp": false,
+      "hiddenAliases": [],
+      "id": "setup:purge",
+      "pluginAlias": "penguins-eggs",
+      "pluginName": "penguins-eggs",
+      "pluginType": "core",
+      "strict": true,
+      "enableJsonFlag": false,
+      "isESM": true,
+      "relativePath": [
+        "dist",
+        "commands",
+        "setup",
+        "purge.js"
       ]
     }
   },
